add intersection point when slope changes sign
While creating tubes, add an extra point where reference
curve changes sign. Use the intersection point between lines
defined by left and right rectangle points and slopes of the
curve on each side of the referece point.

This relaxes a tube a bit and is need to be able to handle
reference curves with hight slope ratios.